pyproject.toml
[tool.poetry]
name = "django-test"
version = "0.1.0"
description = ""
authors = ["rinatrinat <rinatluske@gmail.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.9"
django = "^4.2.1"
django-extensions = "^3.2.1"
django-filter = "^23.2"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.2.2"
drf-nested-routers = "^0.93.4"


[tool.poetry.group.dev.dependencies]
ipython = "^8.13.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
dj = "src.manage:main"


